Clean up forms modules
