+++ netbox/utilities/forms/__init__.py
@@ -1,5 +1,6 @@
 from .constants import *
 from .fields import *
 from .forms import *
+from .mixins import *
 from .utils import *
 from .widgets import *
